split test_eos_unit into multiple, easier to read, faster to compile files
